backend/backend/pyproject.toml
[project]
name = "coact-backend"
version = "0.1.0"
description = "Coact backend: secure device agent orchestration with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Coact", email = "dev@coact.local"}]
dependencies = [
  "botocore>=1.34",
  "boto3>=1.34",
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "aiosqlite>=0.20",
  "alembic>=1.13",
  "passlib[bcrypt]>=1.7",
  "PyJWT[crypto]>=2.8",
  "redis>=5.0",
  "clickhouse-connect>=0.7",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "email-validator>=2.2",
  "loguru>=0.7",
  "aiokafka>=0.10",
  "prometheus-client>=0.20",
  "croniter>=2.0",
  "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "anyio>=4.4",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-redis",
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
  "mkdocs-include-markdown-plugin>=6.0",
  "mkdocs-section-index>=0.3",
  "mkdocs-mermaid2-plugin>=1.1",
  "pymdown-extensions>=10.8",
  "mike>=2.1",
  "websockets>=12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
strict_optional = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"



